requirements.txt
streamlit
python-dotenv
google
google-generativeai